requirements.txt
django-admin-interface
pillow
asgiref==3.7.2
Django==4.2.7
django-jalali==6.0.1
jdatetime==4.1.1
sqlparse==0.4.4
pytz~=2023.3.post1
djangorestframework_simplejwt
djangorestframework
django-filter
celery[redis]
psycopg2-binary
markdown
drf-spectacular